code

+++ db/migrate/20220619231415_create_product_purchases.rb
@@ -3,7 +3,7 @@ class CreateProductPurchases < ActiveRecord::Migration[6.1]
     create_table :product_purchases do |t|
       t.belongs_to :purchase, null: false, foreign_key: true
       t.belongs_to :product, null: false, foreign_key: true
-      t.integer :cost
+      t.float :cost
 
       t.timestamps
     end
